1-basic/requirements.txt
transformers==4.38.2
sentence-transformers==2.5.1
torch
numpy
pandas
tabulate
scikit-learn
matplotlib
requests
pillow
seaborn